[move][move-linter] implement redundant ref deref rules
