--- external-crates/move/crates/move-compiler/tests/custom_rules/shift_overflow.exp
@@ -1,24 +1,0 @@
-warning[Lint W00001]: Potential overflow detected. The number of bits being shifted exceeds the bit width of the variable being shifted.
-  ┌─ tests/custom_rules/shift_overflow.move:5:20
-  │
-5 │         let _b = x << 64; // Should not raise an issue
-  │                    ^^ The 64 of bits being shifted exceeds the 64 bit width of the variable being shifted.
-  │
-  = This warning can be suppressed with '#[allow(lint(shift_overflow))]' applied to the 'module' or module member ('const', 'fun', or 'struct')
-
-warning[Lint W00001]: Potential overflow detected. The number of bits being shifted exceeds the bit width of the variable being shifted.
-  ┌─ tests/custom_rules/shift_overflow.move:6:20
-  │
-6 │         let _b = x << 65; // Should raise an issue
-  │                    ^^ The 65 of bits being shifted exceeds the 64 bit width of the variable being shifted.
-  │
-  = This warning can be suppressed with '#[allow(lint(shift_overflow))]' applied to the 'module' or module member ('const', 'fun', or 'struct')
-
-warning[Lint W00001]: Potential overflow detected. The number of bits being shifted exceeds the bit width of the variable being shifted.
-  ┌─ tests/custom_rules/shift_overflow.move:7:20
-  │
-7 │         let _b = x >> 66; // Should raise an issue
-  │                    ^^ The 66 of bits being shifted exceeds the 64 bit width of the variable being shifted.
-  │
-  = This warning can be suppressed with '#[allow(lint(shift_overflow))]' applied to the 'module' or module member ('const', 'fun', or 'struct')
-
