+++ external-crates/move/crates/move-compiler/tests/linter/correct_redundant_ref_deref.move
@@ -1,0 +1,10 @@
+module 0x42::M {
+    struct MyResource has copy, drop{
+        value: u64,
+    }
+
+    public fun test_borrow_deref_ref() {
+        let resource = MyResource { value: 10 };
+        let _ref1 = &resource;
+    }
+}
