+++ external-crates/move/crates/move-compiler/tests/linter/incorrect_redundant_ref_deref.move
@@ -1,9 +1,8 @@
 module 0x42::M {
+    struct MyResource has copy, drop{
+        value: u64,
+    }
 
-    #[allow(lint(constant_naming))]
-    const Another_BadName: u64 = 42; // Should trigger a warning
-
-    #[allow(lint(redundant_ref_deref))] 
     public fun test_borrow_deref_ref() {
         let resource = MyResource { value: 10 };
         let _ref2 = &*(&resource);
